backend/requirements.txt
django==3.2.18            # Django 3.2 LTS
djangorestframework==3.15
djangorestframework-simplejwt==5.3
djongo==1.3.6
gunicorn==22.0
python-dotenv==1.0
requests==2.32
dj-database-url
django-cors-headers